# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Real state",
        "description": "API para gestionar propiedades inmobiliarias",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/propiedades": {
            "get": {
                "tags": [
                    "Propiedad"
                ],
                "summary": "Obtener todas las propiedades",
                "operationId": "1428dc672691f51f4a853ddcf6faff0e",
                "responses": {
                    "200": {
                        "description": "Lista de propiedades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Propiedad"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propiedad"
                ],
                "summary": "Crear una nueva propiedad",
                "operationId": "8f14f04fae02e4fe798f67893120cbf3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Propiedad"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Propiedad creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Propiedad"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar propiedad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al guardar propiedad: mensaje del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/propiedades/{id}": {
            "get": {
                "tags": [
                    "Propiedad"
                ],
                "summary": "Obtener una propiedad por ID",
                "operationId": "14c81eabb4a966c2adf109f8a124407e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propiedad encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Propiedad"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propiedad no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propiedad no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Propiedad"
                ],
                "summary": "Actualizar una propiedad por ID",
                "operationId": "2baabea464354f0ae482f92a6abfa270",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Propiedad"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Propiedad actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Propiedad"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propiedad no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propiedad no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar propiedad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al actualizar propiedad: mensaje del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Propiedad"
                ],
                "summary": "Eliminar una propiedad por ID",
                "operationId": "ac46a6152a7d0158329dd1adc227c8b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propiedad eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propiedad eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propiedad no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propiedad no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar propiedad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al eliminar propiedad: mensaje del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Propiedad": {
                "title": "Propiedad",
                "description": "Modelo de Propiedad",
                "required": [
                    "direccion",
                    "ciudad",
                    "precio"
                ],
                "properties": {
                    "direccion": {
                        "type": "string",
                        "example": "Calle Falsa 123"
                    },
                    "ciudad": {
                        "type": "string",
                        "example": "Ciudad Ejemplo"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción de la propiedad"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 100000
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "X-RapidAPI-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-RapidAPI-Key"
            }
        }
    ]
}